Fix wrong unit in method occupiedCapacity of Outpoint

+++ core/src/test/java/type/CellOutputTest.java
@@ -33,8 +33,7 @@ public class CellOutputTest {
                 "0x68d5438ac952d2f584abf879527946a537e82c7f3c1cbf6d8ebf9767437d8e88",
                 "0x59a27ef3ba84f061517d13f42cf44ed020610061",
                 Script.HashType.TYPE));
-    Assertions.assertEquals(
-        BigInteger.valueOf(9300000000L),
+    Assertions.assertEquals(9300000000L,
         cellOutput.occupiedCapacity(
             Numeric.hexStringToByteArray(
                 "0x68d5438ac952d2f584abf879527946a537e82c7f3c1cbf6d8ebf9767437d8e88")));
